fix(apis_metainfo): move method from utils to apis_metainfo.utils
The `create_object_from_uri` uses the Uri model, so it makes more sense
if it is located in the app that also contains this model. Besides,
having the method in the `utils.helpers` module led to circular import
errors.

Closes: #1816
