+++ apis_core/generic/forms/fields.py
@@ -2,7 +2,7 @@ from django.core.exceptions import ValidationError
 from django.forms import ModelChoiceField, MultiValueField, MultiWidget
 from django.utils.translation import gettext as _
 
-from apis_core.utils.helpers import create_object_from_uri
+from apis_core.apis_metainfo.utils import create_object_from_uri
 
 
 class ModelImportChoiceField(ModelChoiceField):
